+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `#my_all?` error when passing a fixed single value.
 - `#my_any?` error when passing a fixed single value.
 - `#my_inject` returning incorrect values.
+- `#my_map` not returning enum if no block given.
 
 ### Removed
 - The question mark from `#my_count?` to be `#my_count` instead."